+++ Assets/Hai/ComboGesture/Scripts/Editor/EditorUI/Effectors/CgeBlendTreeEffector.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Hai.ComboGesture.Scripts.Components;
 using Hai.ComboGesture.Scripts.Editor.Internal;
-using Hai.ComboGesture.Scripts.Editor.Internal.Infra;
+using Hai.ComboGesture.Scripts.Editor.Internal.Processing;
 using UnityEditor.Animations;
 using UnityEngine;
 
@@ -139,7 +139,7 @@ namespace Hai.ComboGesture.Scripts.Editor.EditorUI.Effectors
 
         public List<AnimationClip> AllAnimationsOfSelected()
         {
-            return BlendTreeBeingEdited != null ? PuppetToPuppetManifest.AllAnimationsOf(BlendTreeBeingEdited) : new List<AnimationClip>();
+            return BlendTreeBeingEdited != null ? ManifestFromPuppet.AllAnimationsOf(BlendTreeBeingEdited) : new List<AnimationClip>();
         }
     }
