+++ Assets/Hai/ComboGesture/Scripts/Editor/EditorUI/Layouts/CgeLayoutCommon.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Hai.ComboGesture.Scripts.Components;
 using Hai.ComboGesture.Scripts.Editor.EditorUI.Effectors;
-using Hai.ComboGesture.Scripts.Editor.Internal.Infra;
+using Hai.ComboGesture.Scripts.Editor.Internal.Processing;
 using UnityEditor;
 using UnityEditor.Animations;
 using UnityEngine;
@@ -106,7 +106,7 @@ namespace Hai.ComboGesture.Scripts.Editor.EditorUI.Layouts
             }
             else if (element is BlendTree tree)
             {
-                var animations = PuppetToPuppetManifest.AllAnimationsOf(tree);
+                var animations = ManifestFromPuppet.AllAnimationsOf(tree);
                 if (animations.Count == 0)
                 {
                     InvisibleRankPreservingBox();
